requirements.txt
asgiref==3.5.2
certifi==2022.12.7
channels==4.0.0
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==8.5.3
idna==3.4
mongoengine==0.24.2
Pillow==9.3.0
pymongo==4.3.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
